Add character sheet class that can be saved to disk
Created a CharacterSheet class that can hold all the values of a
character sheet. Updated the GET and POST endpoints to load and save to
a file respectively. Updated the character name to be dynamically filled
in as part of the template, as a testing feature.

I'm having a problem with in-memory persistence. When I try to reset an
object, I still keep the character name value upon reload.